+++ primefaces/src/main/java/org/primefaces/model/map/Marker.java
@@ -35,7 +35,7 @@ public class Marker extends Overlay {
 
     private boolean flat;
 
-    private String icon;
+    private Object icon;
 
     private LatLng latlng;
 
@@ -119,11 +119,11 @@ public class Marker extends Overlay {
         this.flat = flat;
     }
 
-    public String getIcon() {
+    public Object getIcon() {
         return icon;
     }
 
-    public void setIcon(String icon) {
+    public void setIcon(Object icon) {
         this.icon = icon;
     }